App struct that owns both window and renderer
This is necessary because while the renderer does not hold a reference
to the window, it needs the window to out-live it (see the safety note
on wgpu::Instance::create_surface).

That fact, combined with the way that winit's EventLoop::run method is
implemented to take a 'static + FnMut, means that the renderer itself
gets a static lifetime when it's moved into the event loop handler
closure.

The simplest way to deal with this is to have a structure that owns both
the window and the renderer to ensure that they have the same lifetime
and can both be moved into the closure together.