Chapter 1: Application draws a triangle
This is a whole lot of code because setting up a pipeline, even without
textures or other bound resources, requires quite a few steps in wgpu.

This *should* be a pretty direct recreation of the OpenGL code from
Chapter 1 of the gltut book, although I'm not yet experienced enough to
judge that properly.